+++ src/Core/ContentManager/ContentManager.php
@@ -415,7 +415,7 @@ class ContentManager
 
         $this->eventDispatcher->dispatch('spress.after_render_page', new Event\RenderEvent(
             $item,
-            ItemInterface::SNAPSHOT_AFTER_RENDER_BLOCKS,
+            ItemInterface::SNAPSHOT_AFTER_PAGE,
             ItemInterface::SNAPSHOT_PATH_RELATIVE
         ));
     }