+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 CHANGELOG for 2.0.x
 ===================
+## 2.0.2
+* [New] `PluginTester` class has been added to the core for testing plugins easily.
+* [Fixed] Fixed an issue with the content retrieved by "after_render_page" event.
+
 ## 2.0.1 (2016-01-09)
 * [Improved] Normalized the directory separator to '/' irrespective of the operating system.
 * [Fixed] Fixed the file's extension `twig.html` in configuration files.